requirements.txt
Django==4.2.3
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
django-redis==5.2.0